Create Sqli_Scan-sh
SQLiScanner is an automated tool that helps you find SQL injection vulnerabilities on websites. It sends special codes to the website's input fields and looks for any signs of errors that suggest the website may be vulnerable. To make sure it is as effective as possible, SQLiScanner uses a range of techniques, including headers, referers, and parameters.

SQLiScanner is a powerful tool that can scan multiple websites at the same time, saving any vulnerable websites it finds in a file for later. It also comes with a variety of codes and messages that can help it identify vulnerabilities more accurately. Whether you are new to website security or an experienced user, SQLiScanner is easy to use and helps you find vulnerabilities quickly and easily.